feat: new clickable area to add plugins
BREAKING CHANGE: its no longer possible to pass a `defaultPlugin`.
The defaultPlugin was used when an Editor or a CellPlugin (former layout plugin) had empty rows.
This was to make sure that there is always something to edit.
That lead to problems and inconsistency.

It was replaced with a new button that appears in every Cell that can take children.
This is more flexible for editors and reduced complexity quite a bit