feat: rework plugins to always have controls and renderer feat: unify all packages into editor package feat: @react-page/create-plugin-materialui is now deprecated feat: improve resize behaviour
BREAKING CHANGE: packages core, renderer and ui have been merged with editor

BREAKING CHANGE: its currently no longer possible to create a custom editor with custom UI and/or multiples editables.
If you need that please file an issue with your use case. Its possible to bring back both, but we want to gather use cases first.
Custom UI elements are now easier to implement, because of the simpler structure and a lot of useful hooks.