fix(plugin-spacer): resizing is broken fix(plugin-spacer): shows border in read only mode fix(plugin-spacer): spacer has now correct size (without 2 additional pixel from the borde)
the border has been removed and replaced with an outline.
This fixes the problem as well that it was actually 2 pixels taller in height as specified.
this might have a slight inpact on existing page layouts and could be considered a breaking change,
but its too sublte and probably few will notice. Please raise an issue if you notice a problem with that.