fix: htmlrenderer might try to render version 0 content when value is ambiguous
we had a situation where due to unknown reasons, the value was migrated to version 1, but it still had `cells`
of version 0. that would not be a problem if the ...rest of the value wouldn't have been sent to all the rows.

that way the cells of every row was overridden with the legacy value.

spreading in the props from the editable like this makes no sense at all, so i removed that