feat(slate): slate uses now a custom static renderer in readonly mode
this reduces bundle size (about 25kb gzipped) and probably render time (not measured yet) and unclutters the dom a bit.

it also introduces a new property `disableProviderInReadOnly` that will not render the provider in readonly mode.
As a matter of fact, slate is the only plugin we provide that uses a provider. It was necessary to also render the Provider in readonly mode,
but that is no longer true. We might deprecate that in the futur or make disableProviderInReadOnly the default.

BREAKING CHANGE: dom in readonly mode is slightly different

the resulting dom in readonly mode differ a bit from previous versions. (less elements)
This had no impact at all in the example project, but depending on your application
you might need to adjust styling. That is why this is released as a breaking change.