feat(slate): read only slate removes unneeded `<span>` tags
this removes a lot of `<span>`'s that have no puprose. With this change,
the dom-tree is significantly smaller (10-20%) when in readonly mode

this could be technically breaking in rare edge cases, but I decided to not mark it as breaking as 3.0.0 just got released
and is a slight improvement on the focus of the 3.0.0 update.

if you experience problems with this, please open up an issue!