feat: improve list plugin
- custom list plugins can now declare `getStyle`
- list plugin no longer need to know the other list plugins,
instead we rely on a prefix on the name.
Only problem is though that the list item uses the same prefix, so we have to rely on that to stay the same.
That means adding a custom list item would be discouraged, but there is not a real usecase anyway for that.